+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* add SesInstanceUserAccessKeySerial to allow rotation of InstanceUserAccessKey
+
 # 3.15.0
 
 * make asg instance types available via python